requirements.txt
# django==4.1.4
django-hitcount==1.3.5
pandas==1.5.2
Pillow==9.3.0
autopep8==2.0.0
openpyxl==3.0.10
django-jazzmin==2.6.0
django-adminrestrict==3.1
gunicorn
libsass
django-compressor
django-sass-processor